Compressing is now angle-limited
During compressing, the angle between segments is checked and compared
with a threshold. If it is bigger, compressing is cut short.

+++ common/c_planner_v5/DisableDebugLog.m
@@ -1,6 +1,6 @@
 function DisableDebugLog(cfg)
 global DebugConfig
 
-    DebugConfig = bitset(DebugConfig, cfg, 0); 
+    DebugConfig = bitset(DebugConfig, int32(cfg), 0); 
 
 end
